refactor(api): split action and run transport schemas
