+++ src/google_work_agent/api/schemas/actions/__init__.py
@@ -1,0 +1,14 @@
+"""Stable external action transport contracts."""
+
+from .approve_action import ActionCommandResponse, ApproveActionRequestV2
+from .modify_action import ModifyActionRequestV2
+from .prepare_retry_action import PrepareRetryRequestV2
+from .reject_action import RejectActionRequestV2
+
+__all__ = [
+    "ActionCommandResponse",
+    "ApproveActionRequestV2",
+    "ModifyActionRequestV2",
+    "PrepareRetryRequestV2",
+    "RejectActionRequestV2",
+]
