+++ src/google_work_agent/api/schemas/actions/prepare_retry_action.py
@@ -1,0 +1,8 @@
+"""Prepare-retry action wire request."""
+
+from google_work_agent.api.schemas.common import ContractVersionedRequest
+
+
+class PrepareRetryRequestV2(ContractVersionedRequest):
+    command_id: str
+    expected_action_version: int
