+++ src/google_work_agent/api/schemas/runs/__init__.py
@@ -1,0 +1,22 @@
+"""Stable external run transport contracts."""
+
+from .cancel_run import CancelRunRequestV2, RunCommandResponse
+from .confirm_run import ConfirmationResponseV1
+from .get_run import RunSnapshotResponse
+from .get_run_context import RunContextResponse
+from .resolve_recovery import ResolveRecoveryRequestV1
+from .resume_run import ResumeRunRequestV2
+from .start_run import SelectedResourceRefModel, StartRunRequest, StartRunResponseModel
+
+__all__ = [
+    "CancelRunRequestV2",
+    "ConfirmationResponseV1",
+    "ResolveRecoveryRequestV1",
+    "ResumeRunRequestV2",
+    "RunCommandResponse",
+    "RunContextResponse",
+    "RunSnapshotResponse",
+    "SelectedResourceRefModel",
+    "StartRunRequest",
+    "StartRunResponseModel",
+]
