+++ src/google_work_agent/api/schemas/runs/get_run.py
@@ -1,0 +1,8 @@
+"""Get-run wire response."""
+
+from google_work_agent.api.schemas.common import ApiModel
+
+
+class RunSnapshotResponse(ApiModel):
+    snapshot: dict[str, object]
+    api_contract_version: str
