+++ src/google_work_agent/api/schemas/runs/get_run_context.py
@@ -1,0 +1,8 @@
+"""Get-run-context wire response."""
+
+from google_work_agent.api.schemas.common import ApiModel
+
+
+class RunContextResponse(ApiModel):
+    context: dict[str, object] | None
+    api_contract_version: str
