+++ src/google_work_agent/api/schemas/runs/resume_run.py
@@ -1,0 +1,11 @@
+"""Resume-run wire request."""
+
+from typing import Literal
+
+from google_work_agent.api.schemas.common import ContractVersionedRequest
+
+
+class ResumeRunRequestV2(ContractVersionedRequest):
+    command_id: str
+    expected_version: int
+    resume_kind: Literal["REAUTH_COMPLETED", "SAFE_CHECKPOINT_RESUME", "RECOVERY_RECHECK"]
