+++ tests/unit/api/test_schema_ownership.py
@@ -1,0 +1,38 @@
+"""Repository-architecture checks for canonical API schema ownership."""
+
+from pathlib import Path
+
+from google_work_agent.api.schemas.actions import (
+    ApproveActionRequestV2,
+    ModifyActionRequestV2,
+    PrepareRetryRequestV2,
+    RejectActionRequestV2,
+)
+from google_work_agent.api.schemas.runs import (
+    CancelRunRequestV2,
+    ConfirmationResponseV1,
+    ResolveRecoveryRequestV1,
+    ResumeRunRequestV2,
+    StartRunRequest,
+)
+
+
+def test_action_transport_contracts_live_in_operation_modules() -> None:
+    assert ApproveActionRequestV2.__module__.endswith(".actions.approve_action")
+    assert ModifyActionRequestV2.__module__.endswith(".actions.modify_action")
+    assert RejectActionRequestV2.__module__.endswith(".actions.reject_action")
+    assert PrepareRetryRequestV2.__module__.endswith(".actions.prepare_retry_action")
+
+
+def test_run_transport_contracts_live_in_operation_modules() -> None:
+    assert StartRunRequest.__module__.endswith(".runs.start_run")
+    assert ConfirmationResponseV1.__module__.endswith(".runs.confirm_run")
+    assert ResumeRunRequestV2.__module__.endswith(".runs.resume_run")
+    assert CancelRunRequestV2.__module__.endswith(".runs.cancel_run")
+    assert ResolveRecoveryRequestV1.__module__.endswith(".runs.resolve_recovery")
+
+
+def test_broad_action_and_run_schema_modules_are_retired() -> None:
+    api_root = Path(__file__).resolve().parents[3] / "src" / "google_work_agent" / "api" / "schemas"
+    assert not (api_root / "actions.py").exists()
+    assert not (api_root / "runs.py").exists()
